# forum/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planinarski Forum API",
        "description": "API dokumentacija za Planinarski Forum aplikaciju",
        "contact": {
            "email": "admin@planinarskiforum.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/posts/{post}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Lajkuj post",
                "description": "Lajkuje određeni post",
                "operationId": "0a240806a4f9df159ba630013f620c01",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID posta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno lajkovano",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Liked"
                                        },
                                        "liked": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        },
                                        "replies_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Ukloni lajk",
                "description": "Uklanja lajk sa posta",
                "operationId": "70f0665199399ff2856fe4e37234df3d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID posta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno uklonjen lajk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unliked"
                                        },
                                        "liked": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        },
                                        "replies_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post}/like/check": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Proveri lajk",
                "description": "Proverava da li je korisnik lajkovao post",
                "operationId": "be36a84c92f57a979f95b8de48d12720",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID posta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post_id": {
                                            "type": "integer"
                                        },
                                        "is_liked": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}